added testing mode (#425)

+++ pywps/app/Process.py
@@ -267,15 +267,20 @@ class Process(object):
     def clean(self):
         """Clean the process working dir and other temporary files
         """
-        LOGGER.info("Removing temporary working directory: %s" % self.workdir)
-        try:
-            if os.path.isdir(self.workdir):
-                shutil.rmtree(self.workdir)
-            if self._grass_mapset and os.path.isdir(self._grass_mapset):
-                LOGGER.info("Removing temporary GRASS GIS mapset: %s" % self._grass_mapset)
-                shutil.rmtree(self._grass_mapset)
-        except Exception as err:
-            LOGGER.error('Unable to remove directory: %s', err)
+        testmode = config.get_config_value('server', 'testmode')
+        if testmode is True:
+            LOGGER.warning('PyWPS is running in testing mode. Workdir is not removed: {}'.format(
+                self.workdir))
+        else:
+            LOGGER.info("Removing temporary working directory: %s" % self.workdir)
+            try:
+                if os.path.isdir(self.workdir):
+                    shutil.rmtree(self.workdir)
+                if self._grass_mapset and os.path.isdir(self._grass_mapset):
+                    LOGGER.info("Removing temporary GRASS GIS mapset: %s" % self._grass_mapset)
+                    shutil.rmtree(self._grass_mapset)
+            except Exception as err:
+                LOGGER.error('Unable to remove directory: %s', err)
 
     def set_workdir(self, workdir):
         """Set working dir for all inputs and outputs
